refactor: modularize scraper into src modules
